requirements.txt
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.29.0
bcrypt==5.0.0
billiard==4.2.4
celery==5.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.104.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
prometheus_client==0.20.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dateutil==2.8.2
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
redis==5.0.3
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.24.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.3.0
websockets==16.0
